pyproject.toml
[tool.poetry]
name = "md2ppt"
version = "0.1.0"
description = ""
authors = ["makai <1345744510@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
python-pptx = "^0.6.23"
setuptools = "^71.1.0"
cython = "^3.0.10"
markdown2 = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
